Add reading sensor data

+++ src/sensor.backend/sensor.Domain/Models/Reading.cs
@@ -3,7 +3,7 @@
     public class Reading
     {
         public string? Id { get; init; }
-        public int Value { get; init; }
+        public float Value { get; init; }
         public DateTime Timestamp { get; init; }
         public ReadingMetadata Metadata { get; init; }
     }
